refactoring: move out sendMessage
