+++ lib/sync/matrix/consts.dart
@@ -1,0 +1,2 @@
+const configNotFound = 'Could not find Matrix Config';
+const syncMessageType = 'com.lotti.sync.message';
