+++ lib/sync/matrix/stats.dart
@@ -1,0 +1,9 @@
+class MatrixStats {
+  MatrixStats({
+    required this.sentCount,
+    required this.messageCounts,
+  });
+
+  int sentCount;
+  Map<String, int> messageCounts;
+}
