+++ lib/widgets/sync/matrix/incoming_stats.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/sync/matrix/matrix_service.dart';
+import 'package:lotti/sync/matrix/stats.dart';
 
 class IncomingStats extends StatelessWidget {
   const IncomingStats({super.key});
